Change: Increase limit for max_num_autosaves from 255 to UINT32_MAX
This makes it easier to compare how AIs behave from outside of OpenTTD, for
example by using https://github.com/TrueBrain/OpenTTD-savegame-reader to open
many savegame files and plotting how properties, like money, change every
month, or even every day, over a long period of time.

This is helpful both for the development of AIs to improve gameplay, and when
wanting to use OpenTTD for research purposes.

+++ src/settings_type.h
@@ -155,7 +155,7 @@ struct GUISettings {
 	bool   autosave_on_exit;                 ///< save an autosave when you quit the game, but do not ask "Do you really want to quit?"
 	bool   autosave_on_network_disconnect;   ///< save an autosave when you get disconnected from a network game with an error?
 	uint8_t  date_format_in_default_names;     ///< should the default savegame/screenshot name use long dates (31th Dec 2008), short dates (31-12-2008) or ISO dates (2008-12-31)
-	byte   max_num_autosaves;                ///< controls how many autosavegames are made before the game starts to overwrite (names them 0 to max_num_autosaves - 1)
+	uint32_t max_num_autosaves;              ///< controls how many autosavegames are made before the game starts to overwrite (names them 0 to max_num_autosaves - 1)
 	bool   population_in_label;              ///< show the population of a town in its label?
 	uint8_t  right_mouse_btn_emulation;        ///< should we emulate right mouse clicking?
 	uint8_t  scrollwheel_scrolling;            ///< scrolling using the scroll wheel?
